requirements.txt
aniso8601==10.0.0
bcrypt==4.2.1
blinker==1.9.0
click==8.1.8
dnspython==2.7.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-DotEnv==0.1.2
Flask-RESTful==0.3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mongoengine==0.29.1
pymongo==4.11.1
python-dotenv==1.0.1
pytz==2025.1
six==1.17.0
Werkzeug==3.1.3
